combine nodejs files
