+++ bindings/nodejs/lib/types/wallet/create-delegation-transaction.ts
@@ -1,12 +1,11 @@
 // Copyright 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-    Wallet,
-    DelegationId,
-    PreparedCreateDelegationTransactionData,
-} from '../..';
+import { Wallet, DelegationId } from '../..';
 import { PreparedTransaction } from './prepared-transaction';
+import { Type } from 'class-transformer';
+import { TransactionWithMetadata } from './transaction';
+import { PreparedTransactionData } from '../client/prepared-transaction-data';
 
 /*
  * Represents prepared data for issuing a transaction to create a delegation.
@@ -34,4 +33,22 @@ export class PreparedCreateDelegationTransaction extends PreparedTransaction {
     public delegationId(): DelegationId {
         return this._delegationId;
     }
+}
+
+/** The result of preparing an operation to create a native token. */
+export class PreparedCreateDelegationTransactionData {
+    /** The id of the delegation */
+    delegationId!: DelegationId;
+    /** The prepared transaction which will create the delegation */
+    @Type(() => PreparedTransactionData)
+    transaction!: PreparedTransactionData;
+}
+
+/** The result of an operation to create a native token. */
+export class CreateDelegationTransaction {
+    /** The id of the delegation */
+    delegationId!: DelegationId;
+    /** The transaction which created the delegation */
+    @Type(() => TransactionWithMetadata)
+    transaction!: TransactionWithMetadata;
 }
