+++ bindings/nodejs/lib/types/wallet/prepared-create-token-transaction.ts
@@ -1,8 +1,11 @@
 // Copyright 2021-2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { Wallet, PreparedCreateNativeTokenTransactionData } from '../..';
+import { Wallet } from '../..';
 import { PreparedTransaction } from './prepared-transaction';
+import { Type } from 'class-transformer';
+import { TransactionWithMetadata } from './transaction';
+import { PreparedTransactionData } from '../client/prepared-transaction-data';
 
 /*
  * The class PreparedCreateNativeTokenTransaction represents prepared data for issuing a transaction to create a native token.
@@ -32,4 +35,22 @@ export class PreparedCreateNativeTokenTransaction extends PreparedTransaction {
     public tokenId(): string {
         return this._tokenId;
     }
+}
+
+/** The result of preparing an operation to create a native token. */
+export class PreparedCreateNativeTokenTransactionData {
+    /** The token id of the minted token */
+    tokenId!: string;
+    /** The prepared transaction which will mint the token */
+    @Type(() => PreparedTransactionData)
+    transaction!: PreparedTransactionData;
+}
+
+/** The result of an operation to create a native token. */
+export class CreateNativeTokenTransaction {
+    /** The token id of the minted token */
+    tokenId!: string;
+    /** The transaction which minted the token */
+    @Type(() => TransactionWithMetadata)
+    transaction!: TransactionWithMetadata;
 }
